+++ goodsending/src/main/java/com/goodsending/product/dto/response/ProductRankingDto.java
@@ -1,8 +1,11 @@
 package com.goodsending.product.dto.response;
 
+import com.goodsending.product.entity.Product;
+import com.goodsending.product.entity.ProductImage;
 import com.goodsending.product.type.ProductStatus;
 import com.querydsl.core.annotations.QueryProjection;
 import java.time.LocalDateTime;
+import lombok.Builder;
 import lombok.Getter;
 
 @Getter
@@ -16,6 +19,7 @@ public class ProductRankingDto {
   private ProductStatus status;
   private String thumbnailUrl;
 
+  @Builder
   @QueryProjection
   public ProductRankingDto(Long productId, String name, int price, LocalDateTime startDateTime,
       LocalDateTime maxEndDateTime, ProductStatus status, String thumbnailUrl) {
@@ -26,5 +30,17 @@ public class ProductRankingDto {
     this.maxEndDateTime = maxEndDateTime;
     this.status = status;
     this.thumbnailUrl = thumbnailUrl;
+  }
+
+  public static ProductRankingDto of(Product product, ProductImage productImage) {
+    return ProductRankingDto.builder()
+        .productId(product.getId())
+        .name(product.getName())
+        .price(product.getPrice())
+        .startDateTime(product.getStartDateTime())
+        .maxEndDateTime(product.getMaxEndDateTime())
+        .status(product.getStatus())
+        .thumbnailUrl(productImage.getUrl())
+        .build();
   }
 }
