+++ goodsending/src/main/java/com/goodsending/product/repository/ProductCustomRepositoryImpl.java
@@ -10,8 +10,6 @@ import com.goodsending.product.dto.response.QProductRankingDto;
 import com.goodsending.product.dto.response.QProductSummaryDto;
 import com.goodsending.product.entity.Product;
 import com.goodsending.product.type.ProductStatus;
-import com.goodsending.productlike.dto.ProductRankingDto;
-import com.goodsending.productlike.dto.QProductRankingDto;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.types.dsl.CaseBuilder;
@@ -201,22 +199,6 @@ public class ProductCustomRepositoryImpl implements ProductCustomRepository {
       closedBuilder.and(product.status.eq(ProductStatus.ENDED));
     }
     return closedBuilder;
-  }
-
-  @Override
-  public ProductRankingDto findRankingDtoById(Long productId) {
-    return jpaQueryFactory.select(new QProductRankingDto(
-            product.id,
-            product.name,
-            product.price,
-            product.startDateTime,
-            product.maxEndDateTime,
-            product.status,
-            productImage.url))
-        .from(product)
-        .leftJoin(productImage).on(productImage.product.eq(product))
-        .where(productImageEq().and(product.id.eq(productId)))
-        .fetchOne();
   }
 }
