+++ goodsending/src/main/java/com/goodsending/productlike/repository/LikeCountRankingRepository.java
@@ -1,6 +1,6 @@
 package com.goodsending.productlike.repository;
 import com.goodsending.global.redis.RedisRankingRepository;
-import com.goodsending.productlike.dto.ProductRankingDto;
+import com.goodsending.product.dto.response.ProductRankingDto;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Repository;
